Define a more relaxed metadata schema as well
Closes #12 by offering the choice of two top level schemas,

* `ebu-tt-metadata.xsd` defines metadata types for use in various TTML elements applying strict
validation to _should_ specifications in Tech3390 or other EBU-TT specifications, for
example restricting the use of `ebuttm:facet` to `ttm:metadata` elements that can be children of
content elements, and the use of `ebuttm:documentFacet` to the metadata child of the head
element.
* `ebu-tt-metadata-relaxed.xsd` defines all the metadata elements specified in Tech3390
and allows any of them to be used on any `ttm:metadata` element in any order, applying no
further constraints except that the deprecated element `ebuttm:documentMetadata` is not
permitted.

The README is updated in this commit to reflect the above.

To achieve this, the element definitions were refactored into a separate xsd that is included by both the top level schema files, so element definitions need only be changed in a single location.